Move usage stats into a component
This will allow us to pull a big chunk of the state out of the parent
component, so it can focus on other stuff.
