+++ apps/client/lib/client_web/controllers/water_log_kiosk_live.ex
@@ -1,11 +1,8 @@
 defmodule ClientWeb.WaterLogKioskLive do
   require Logger
   use Phoenix.LiveView
-
-  alias Client.{
-    Util,
-    WaterLogs
-  }
+  alias Client.WaterLogs
+  alias ClientWeb.WaterLogs.Usage
 
   def render(assigns) do
     ~L"""
@@ -21,14 +18,8 @@ defmodule ClientWeb.WaterLogKioskLive do
         <% end %>
       </div>
 
-
       <div class="mt-5 text-xl">
-        <div>
-          Dispensed today: <%= Util.format_number(@today_ml) %> ml
-        </div>
-        <div class="mt-3">
-          Total dispensed: <%= Util.format_number(@total_ml) %> ml
-        </div>
+        <%= live_component(@socket, Usage, usage_state(@socket)) %>
       </div>
     </div>
     """
@@ -47,9 +38,7 @@ defmodule ClientWeb.WaterLogKioskLive do
       ml: 0,
       saving: false,
       log_id: log_id,
-      log: log,
-      today_ml: amount_dispensed_today(log),
-      total_ml: total_amount_dispensed(log)
+      log: log
     }
 
     {:ok, assign(socket, assigns)}
@@ -64,33 +53,13 @@ defmodule ClientWeb.WaterLogKioskLive do
   end
 
   def handle_info(:saved, socket) do
-    assigns = %{
-      saving: false,
-      ml: 0,
-      total_ml: total_amount_dispensed(socket.assigns[:log]),
-      today_ml: amount_dispensed_today(socket.assigns[:log])
-    }
+    send_update(Usage, usage_state(socket))
 
+    assigns = %{saving: false, ml: 0}
     {:noreply, assign(socket, assigns)}
   end
 
-  defp amount_dispensed_today(log) do
-    beginning_of_today = now() |> beginning_of_day()
-    WaterLogs.get_amount_dispensed(log.id, start_at: beginning_of_today)
-  end
-
-  def total_amount_dispensed(log) do
-    start_at = DateTime.from_naive!(log.inserted_at, "Etc/UTC")
-    WaterLogs.get_amount_dispensed(log.id, start_at: start_at)
-  end
-
-  defp now do
-    :client
-    |> Application.fetch_env!(:default_timezone)
-    |> DateTime.now!()
-  end
-
-  defp beginning_of_day(datetime) do
-    %{datetime | hour: 0, minute: 0, second: 0, microsecond: {0, 0}}
+  defp usage_state(socket) do
+    [id: :usage, log: socket.assigns[:log]]
   end
 end
